requirements.txt
Django
gunicorn
django-allauth
dj-rest-auth
djangorestframework
whitenoise
dj-database-url
django-heroku
django-cors-headers
pytest
pytest-django